+++ internal/datasource/cardano_db_sync/models/pool_retire.go
@@ -3,7 +3,7 @@ package models
 type PoolRetire struct {
 	Id            int64  `gorm:"column:id"`
 	HashId        int64  `gorm:"column:hash_id"` // pool_hash(id)
-	CertIndex     int64  `gorm:"column:cert_index"`
+	CertIndex     int32  `gorm:"column:cert_index"`
 	AnnounceTxnId int64  `gorm:"column:announced_tx_id"` // tx(id)
 	RetiringEpoch uint64 `gorm:"column:retiring_epoch"`
 }
